+++ src/badge-buddy-api-dto/auth/token-get-request/token-get-request.dto.spec.ts
@@ -1,4 +1,3 @@
-
 import { describe, it, expect } from '@jest/globals';
 import { TokenGetRequestDto } from './token-get-request.dto';