+++ src/badge-buddy-api-dto/community-events-active-discord/community-events-active-discord-by-voice-channel-get-request/community-events-active-discord-by-voice-channel-get-request.dto.spec.ts
@@ -1,0 +1,10 @@
+import { CommunityEventsActiveDiscordByVoiceChannelGetRequestDto } from './community-events-active-discord-by-voice-channel-get-request.dto';
+import { describe, it, expect } from '@jest/globals';
+
+describe('CommunityEventsActiveByVoiceChannelGetRequestDto', () => {
+  it('should be defined', () => {
+    expect(
+      new CommunityEventsActiveDiscordByVoiceChannelGetRequestDto(),
+    ).toBeDefined();
+  });
+});