+++ src/badge-buddy-api-dto/community-events-active-discord/community-events-active-discord-get-response/community-events-active-discord-get-response.dto.ts
@@ -1,0 +1,13 @@
+import { IsArray } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+import { CommunityEventActiveDiscordDto } from '../community-event-active-discord.dto';
+
+export class CommunityEventsActiveDiscordGetResponseDto {
+  @ApiProperty({
+    required: true,
+    description: 'List of active events',
+    type: [CommunityEventActiveDiscordDto],
+  })
+  @IsArray()
+  events: CommunityEventActiveDiscordDto[];
+}