+++ src/badge-buddy-api-dto/community-events/discord-community-event-patch-response/discord-community-event-patch-response.dto.spec.ts
@@ -1,8 +1,0 @@
-import { DiscordCommunityEventPatchResponseDto } from './discord-community-event-patch-response.dto';
-import { describe, it, expect } from '@jest/globals';
-
-describe('DiscordCommunityEventPatchResponseDto', () => {
-  it('should be defined', () => {
-    expect(new DiscordCommunityEventPatchResponseDto()).toBeDefined();
-  });
-});