+++ src/badge-buddy-api-dto/community-events/discord-community-event-post-request/discord-community-event-post-request.dto.spec.ts
@@ -1,8 +1,0 @@
-import { DiscordCommunityEventPostRequestDto } from './discord-community-event-post-request.dto';
-import { describe, it, expect } from '@jest/globals';
-
-describe('DiscordCommunityEventPostRequestDto', () => {
-  it('should be defined', () => {
-    expect(new DiscordCommunityEventPostRequestDto()).toBeDefined();
-  });
-});