+++ src/badge-buddy-api-dto/community-events/discord-community-events-active-by-guild-and-organizer-get-request/discord-community-events-active-by-guild-and-organizer-get-request.dto.spec.ts
@@ -1,8 +1,0 @@
-import { DiscordCommunityEventsActiveByGuildAndOrganizerGetRequestDto } from './discord-community-events-active-by-guild-and-organizer-get-request.dto';
-import { describe, it, expect } from '@jest/globals';
-
-describe('CommunityEventsActiveByGuildAndOrganizerGetRequestDto', () => {
-  it('should be defined', () => {
-    expect(new DiscordCommunityEventsActiveByGuildAndOrganizerGetRequestDto()).toBeDefined();
-  });
-});