+++ src/badge-buddy-api-dto/community-events/discord-community-events-active-by-organizer-get-request/discord-community-events-active-by-organizer-get-request.dto.spec.ts
@@ -1,8 +1,0 @@
-import { DiscordCommunityEventsActiveByOrganizerGetRequestDto } from './discord-community-events-active-by-organizer-get-request.dto';
-import { describe, it, expect } from '@jest/globals';
-
-describe('CommunityEventsActiveByOrganizerGetRequestDto', () => {
-  it('should be defined', () => {
-    expect(new DiscordCommunityEventsActiveByOrganizerGetRequestDto()).toBeDefined();
-  });
-});