+++ src/badge-buddy-api-dto/community-events/discord-community-events-active-by-voice-channel-get-request/discord-community-events-active-by-voice-channel-get-request.dto.spec.ts
@@ -1,8 +1,0 @@
-import { DiscordCommunityEventsActiveByVoiceChannelGetRequestDto } from './discord-community-events-active-by-voice-channel-get-request.dto';
-import { describe, it, expect } from '@jest/globals';
-
-describe('CommunityEventsActiveByVoiceChannelGetRequestDto', () => {
-  it('should be defined', () => {
-    expect(new DiscordCommunityEventsActiveByVoiceChannelGetRequestDto()).toBeDefined();
-  });
-});