+++ src/badge-buddy-api-dto/poaps/poaps-store-discord-post-response/poaps-store-discord-post-response.dto.ts
@@ -1,0 +1,11 @@
+import { IsArray } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class PoapsStoreDiscordPostResponseDto {
+  @IsArray()
+  @ApiProperty({
+    description: 'List of POAPs that were not distributed to participants',
+    example: ['https://poap.xyz/claim/123'],
+  })
+  poapsAvailable: string[];
+}