+++ src/common-typeorm/scripts/mysql.setup.sql
@@ -1,5 +1,4 @@
-CREATE DATABASE badge_buddy;
-DROP DATABASE badge_buddy;
+USE badge_buddy;
 
 -- Auth
 
@@ -72,7 +71,8 @@ CREATE TABLE community_events
   end_date DATETIME NOT NULL,
   image TEXT,
   website TEXT,
-  poap_event_id INT -- https://documentation.poap.tech/reference/geteventsid
+  poap_event_id INT, -- https://documentation.poap.tech/reference/geteventsid
+  disbanded_date DATETIME
 );
 
 CREATE TABLE community_events_discord
@@ -108,9 +108,10 @@ CREATE TABLE poap_claims_discord
   claimed_on DATETIME,
   expires_on DATETIME,
   FOREIGN KEY (poap_claim_id) REFERENCES poap_claims (id),
-  FOREIGN KEY (assigned_discord_user_id) REFERENCES discord_users (id),
-  FOREIGN KEY (assigned_discord_user_sid) REFERENCES discord_users (user_sid)
+  INDEX assigned_discord_user_sid_idx (assigned_discord_user_sid),
+  INDEX assigned_discord_user_id_idx (assigned_discord_user_id)
 );
+
 
 -- Community Participants
 
@@ -122,7 +123,6 @@ CREATE TABLE community_events_participants_discord
   end_date DATETIME,
   participation_length INT UNSIGNED, -- in seconds
   FOREIGN KEY (community_event_id) REFERENCES community_events_discord (community_event_id),
-  FOREIGN KEY (discord_user_sid) REFERENCES discord_users (user_sid),
-  FOREIGN KEY (discord_user_sid) REFERENCES poap_claims_discord (assigned_discord_user_sid),
-  PRIMARY KEY (community_event_id, discord_user_sid)
+  PRIMARY KEY (community_event_id, discord_user_sid),
+  INDEX discord_user_sid_idx (discord_user_sid)
 );